AI_takeaki_final/requirements.txt
streamlit
langchain
faiss-cpu
openai
python-dotenv
docx2txt
langchain-community
protobuf<=3.20.3